move decoder.py, mstlstm.py, parse.py into bist dir
